Lua API angelegt und implementiert LED_Array struct angelegt, size attribut in uint16_t konvertiert HTTP_API angelegt, functions prototyped

+++ include/CustomTypes/Main_Config.h
@@ -8,7 +8,7 @@
 #include "Arduino.h"
 
 struct MainConfig {
-    int num_leds;
+    uint16_t num_leds;
     uint8_t max_brightness;
     uint8_t current_brightness;
     String last_id;
